Fix forge 1.20.4+ installs.
Fixes trying to download the -client jars. Which are generated by the installation process, denoted by the artifact having no download URL in the manifest. The existing code here was filtering the wrong list.

If this becomes a problem, we can just run 2 library install passes, one prior to the processors, for their depencencies. And one after to download any runtime libraries, the -client jar in this scenario would pass hash validation and have its download skipped.