Workaround for ignoring forge processor outputs.
We have seen zlib-ng starting to gain popularity. This is causing issues as it generates different deflate blobs compared to regular zlib. Forge installers rely on the output semantics of zlib, as it builds jars.

Ideally Forge should probably just not compress these zips, as this will be an ongoing issue. But this is nothing we have control over here.